server/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.26.0
sse-starlette>=2.0.0
chromadb>=0.4.0
tavily-python>=0.3.0
openai-agents>=0.0.5
requests>=2.31.0
